Fix bugs + add comprehensive test suite (91/91 passing)
- Fix i64_scatter dtype mismatch (int64 vs int32 in scatter_indices)
- Fix CLI bench import path (sys.path for benchmarks module)
- Fix pyproject.toml: remove unused bitsandbytes dep
- Fix test_async_engine: add pytest.mark.asyncio markers for strict mode
- Add test_model.py: INLDynamics, MuGuidedAttention, MuGuidedTokenRoutedMLP, ComplexityDecoderLayer, full model forward
- Add test_loader.py: config parsing, from_json, weight loading, tied embeddings
- Add test_api_server.py: health, models, completions, chat, concurrent requests
- Add test_tensor_parallel.py: ColumnParallel, RowParallel, shard_expert_weights, all_reduce
- Add test_utils.py: sampling, quantization, chat template, RMSNorm, rotary embeddings

+++ vllm_i64/kernels/i64_ops.py
@@ -72,10 +72,10 @@ def i64_scatter(
     # Build reverse mapping
     scatter_indices = torch.zeros_like(sorted_indices)
     scatter_indices[sorted_indices] = torch.arange(
-        len(sorted_indices), dtype=torch.int32, device=hidden_states.device
+        len(sorted_indices), dtype=sorted_indices.dtype, device=hidden_states.device
     )
 
-    return scattered, scatter_indices, expert_offsets, expert_counts
+    return scattered, scatter_indices.to(torch.int32), expert_offsets, expert_counts
 
 
 def i64_gather(
